arm64/vmm: Use a capability to represent the guest PC in purecap kernels
hypctx->tf.tf_elr is loaded into ELR prior to dropping back to EL1.

+++ usr.sbin/bhyve/aarch64/vmexit.c
@@ -79,7 +79,7 @@ vmexit_inst_emul(struct vmctx *ctx __unused, struct vcpu *vcpu,
 
 fail:
 	fprintf(stderr, "Failed to emulate instruction ");
-	FPRINTLN(stderr, "at 0x%lx", vme->pc);
+	FPRINTLN(stderr, "at 0x%lx", (uint64_t)vme->pc);
 	return (VMEXIT_ABORT);
 }
